Subsections with prereqs now visible in outline
WL-1317

Previously, the course outline would only show subsections that had
prerequisites if the prerequisite was already met. It would give no
indication that the subsection existed before the prequisite was met.
The course outline will now show subsections with indications that that
section (a) has prerequisites and (b) whether the prerequisite has been
met.